these should have remained ints
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@197686 c90b9560-bf6c-de11-be94-00142212c4b1

+++ ext/xmlreader/php_xmlreader.c
@@ -944,7 +944,7 @@ Sets the string that the the XMLReader will parse. */
 PHP_METHOD(xmlreader, XML)
 {
 	zval *id;
-	long source_len = 0, encoding_len = 0;
+	int source_len = 0, encoding_len = 0;
 	long options = 0;
 	xmlreader_object *intern;
 	char *source, *uri = NULL, *encoding = NULL;
